+++ contrib/spacy/test_spacy.py
@@ -3,7 +3,7 @@ import time
 from spacy_huggingface_hub import push
 
 from huggingface_hub import delete_repo, hf_hub_download, model_info
-from huggingface_hub.utils import HfHubHTTPError
+from huggingface_hub.errors import HfHubHTTPError
 
 from ..utils import production_endpoint
