+++ src/huggingface_hub/hf_api.py
@@ -102,19 +102,21 @@ from .constants import (
     DiscussionStatusFilter,
     DiscussionTypeFilter,
 )
+from .errors import (
+    BadRequestError,
+    EntryNotFoundError,
+    GatedRepoError,
+    HfHubHTTPError,
+    RepositoryNotFoundError,
+    RevisionNotFoundError,
+)
 from .file_download import HfFileMetadata, get_hf_file_metadata, hf_hub_url
 from .repocard_data import DatasetCardData, ModelCardData, SpaceCardData
 from .utils import (
     DEFAULT_IGNORE_PATTERNS,
-    BadRequestError,
-    EntryNotFoundError,
-    GatedRepoError,
     HfFolder,  # noqa: F401 # kept for backward compatibility
-    HfHubHTTPError,
     LocalTokenNotFoundError,
     NotASafetensorsRepoError,
-    RepositoryNotFoundError,
-    RevisionNotFoundError,
     SafetensorsFileMetadata,
     SafetensorsParsingError,
     SafetensorsRepoMetadata,
