+++ tests/test_cache_layout.py
@@ -4,8 +4,8 @@ import unittest
 from io import BytesIO
 
 from huggingface_hub import HfApi, hf_hub_download, snapshot_download
+from huggingface_hub.errors import EntryNotFoundError
 from huggingface_hub.utils import SoftTemporaryDirectory, logging
-from huggingface_hub.utils._errors import EntryNotFoundError
 
 from .testing_constants import ENDPOINT_STAGING, TOKEN
 from .testing_utils import (
