+++ tests/test_hf_api.py
@@ -49,6 +49,13 @@ from huggingface_hub.constants import (
     REPO_TYPE_SPACE,
     SPACES_SDK_TYPES,
 )
+from huggingface_hub.errors import (
+    BadRequestError,
+    EntryNotFoundError,
+    HfHubHTTPError,
+    RepositoryNotFoundError,
+    RevisionNotFoundError,
+)
 from huggingface_hub.file_download import hf_hub_download
 from huggingface_hub.hf_api import (
     AccessRequest,
@@ -70,12 +77,7 @@ from huggingface_hub.hf_api import (
 )
 from huggingface_hub.repocard_data import DatasetCardData, ModelCardData
 from huggingface_hub.utils import (
-    BadRequestError,
-    EntryNotFoundError,
-    HfHubHTTPError,
     NotASafetensorsRepoError,
-    RepositoryNotFoundError,
-    RevisionNotFoundError,
     SafetensorsFileMetadata,
     SafetensorsParsingError,
     SafetensorsRepoMetadata,
