+++ tests/test_repocard.py
@@ -37,11 +37,12 @@ from huggingface_hub import (
     metadata_update,
 )
 from huggingface_hub.constants import REPOCARD_NAME
+from huggingface_hub.errors import EntryNotFoundError
 from huggingface_hub.file_download import hf_hub_download
 from huggingface_hub.hf_api import HfApi
 from huggingface_hub.repocard import REGEX_YAML_BLOCK
 from huggingface_hub.repocard_data import CardData
-from huggingface_hub.utils import EntryNotFoundError, SoftTemporaryDirectory, is_jinja_available
+from huggingface_hub.utils import SoftTemporaryDirectory, is_jinja_available
 
 from .testing_constants import (
     ENDPOINT_STAGING,
